Add files via upload
Why Summarization?We all interact with applications which uses text summarization.Many of those applications are for the platform which publishesarticles on daily news, entertainment, sports. With our busyschedule, we prefer to read the summary of those article before wedecide to jump in for reading entire article. Reading a summary helpus to identify the interest area, gives a brief context of the story.